refactor(test): split test_inspector.cpp by concern into four sibling TUs
test_inspector.cpp was a 4,135-line Catch2 test file for the inspect/
subsystem. Carve four concern-scoped clusters into sibling test TUs,
moving whole TEST_CASE blocks byte-verbatim — same names, tags, and
assertions. Only the file/binary each case lives in changed.

New files:
  test_inspector_gpu_passes.cpp      (11 cases) — Phase 6.1 render-pass
                                                  attribution viewer
  test_inspector_source_jump.cpp     ( 9 cases) — Phase 5.1/5.3 source-jump
  test_inspector_drift_reconcile.cpp (17 cases) — Phase 2 drift drawer +
                                                  Phase 5.2 reconcile tab
  test_inspector_atlas_viewer.cpp    ( 7 cases) — Phase 6.2 atlas viewer

test_inspector.cpp: 4,135 -> 2,910 lines (-1,225). Each new file is
registered in test/CMakeLists.txt mirroring pulp-test-inspector's exact
pattern — own executable + catch_discover_tests, same libraries, same
PULP_INSPECTOR_NO_LAUNCH env guard. Three small file-local helpers
(ScopedEnv, make_key, count_text_containing) were duplicated verbatim
into the new files that need them so each cluster is self-contained.

Pre-split: 150 cases / 816 assertions in one binary.
Post-split: 150 cases / 816 assertions across five binaries
            (106 + 11 + 9 + 17 + 7). Counts match exactly.
