+++ test/test_inspector.cpp
@@ -2908,1228 +2908,3 @@ TEST_CASE("LiveConstant.reset rolls a value back to its default",
     REQUIRE_THAT(registry.get("test_reset"),
                  Catch::Matchers::WithinAbs(2.0, 0.001));
 }
-
-// ── Phase 6.1 — per-pass GPU/render attribution viewer ──────────────────────
-//
-// Spec: planning/2026-05-19-inspector-phase6-gpu-perf-spike.md § Phase 6.1.
-// The viewer surfaces per-pass render cost over a rolling 60-frame window
-// using RenderPassManager's existing CPU-time PassStats. True GPU
-// timestamps are deferred to Phase 6.5 — these tests assert only what the
-// RenderPassManager can actually report today.
-
-namespace {
-
-using pulp::render::RenderPassManager;
-using pulp::render::RenderPassType;
-
-// Drive one synthetic render frame through the manager: begin a frame,
-// emit each (type, time_ms, draw_calls) pass, end the frame.
-void render_synthetic_frame(
-    RenderPassManager& rpm,
-    const std::vector<std::tuple<RenderPassType, float, int>>& passes) {
-    rpm.begin_frame();
-    for (auto& [type, ms, dc] : passes) {
-        rpm.begin_pass(type);
-        rpm.end_pass(ms, dc);
-    }
-    rpm.end_frame();
-}
-
-// Count RecordingCanvas fill_text commands whose text contains `needle`.
-int count_text_containing(const pulp::canvas::RecordingCanvas& canvas,
-                          std::string_view needle) {
-    int n = 0;
-    for (const auto& cmd : canvas.commands()) {
-        if (cmd.type == pulp::canvas::DrawCommand::Type::fill_text &&
-            cmd.text.find(needle) != std::string::npos)
-            ++n;
-    }
-    return n;
-}
-
-} // namespace
-
-TEST_CASE("InspectorOverlay Phase 6.1: capture_pass_frame is a no-op without an RPM",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-
-    REQUIRE_FALSE(overlay.capture_pass_frame());
-    REQUIRE(overlay.pass_frames_captured() == 0);
-    // Every pass entry reports zero history.
-    for (const auto& a : overlay.pass_attribution())
-        REQUIRE(a.samples == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: capture accumulates per-pass history",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-
-    RenderPassManager rpm;
-    overlay.set_render_pass_manager(&rpm);
-
-    // Frame 1: background + content passes.
-    render_synthetic_frame(rpm, {
-        {RenderPassType::background, 1.0f, 3},
-        {RenderPassType::content,    4.0f, 12},
-    });
-    REQUIRE(overlay.capture_pass_frame());
-    REQUIRE(overlay.pass_frames_captured() == 1);
-
-    // Frame 2: content gets heavier, an overlay pass appears.
-    render_synthetic_frame(rpm, {
-        {RenderPassType::background, 1.0f, 3},
-        {RenderPassType::content,    8.0f, 20},
-        {RenderPassType::overlay,    2.0f, 5},
-    });
-    REQUIRE(overlay.capture_pass_frame());
-    REQUIRE(overlay.pass_frames_captured() == 2);
-
-    auto attrib = overlay.pass_attribution();
-    REQUIRE(attrib.size() == 5);  // one entry per RenderPassType.
-
-    // background: two samples, steady at 1.0ms.
-    const auto& bg = attrib[static_cast<int>(RenderPassType::background)];
-    REQUIRE(bg.samples == 2);
-    REQUIRE(bg.present);
-    REQUIRE_THAT(bg.last_cpu_ms, Catch::Matchers::WithinAbs(1.0, 0.001));
-    REQUIRE_THAT(bg.avg_cpu_ms, Catch::Matchers::WithinAbs(1.0, 0.001));
-    REQUIRE(bg.last_draw_calls == 3);
-
-    // content: avg of 4 and 8 = 6, peak 8, last 8.
-    const auto& content = attrib[static_cast<int>(RenderPassType::content)];
-    REQUIRE(content.samples == 2);
-    REQUIRE_THAT(content.last_cpu_ms, Catch::Matchers::WithinAbs(8.0, 0.001));
-    REQUIRE_THAT(content.avg_cpu_ms, Catch::Matchers::WithinAbs(6.0, 0.001));
-    REQUIRE_THAT(content.peak_cpu_ms, Catch::Matchers::WithinAbs(8.0, 0.001));
-    REQUIRE(content.peak_draw_calls == 20);
-
-    // overlay: only one sample (absent from frame 1).
-    const auto& overlay_pass = attrib[static_cast<int>(RenderPassType::overlay)];
-    REQUIRE(overlay_pass.samples == 1);
-    REQUIRE(overlay_pass.present);
-
-    // effects + post never rendered — no history, not present.
-    const auto& effects = attrib[static_cast<int>(RenderPassType::effects)];
-    REQUIRE(effects.samples == 0);
-    REQUIRE_FALSE(effects.present);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: capture de-dups within a single frame",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-
-    RenderPassManager rpm;
-    overlay.set_render_pass_manager(&rpm);
-
-    render_synthetic_frame(rpm, {{RenderPassType::content, 5.0f, 10}});
-
-    // First capture records the frame; a second capture of the SAME
-    // frame (no begin_frame between) must be a no-op so paint() running
-    // multiple times per frame doesn't inflate history.
-    REQUIRE(overlay.capture_pass_frame());
-    REQUIRE_FALSE(overlay.capture_pass_frame());
-    REQUIRE(overlay.pass_frames_captured() == 1);
-
-    auto attrib = overlay.pass_attribution();
-    REQUIRE(attrib[static_cast<int>(RenderPassType::content)].samples == 1);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: same pass type twice in a frame sums",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-
-    RenderPassManager rpm;
-    overlay.set_render_pass_manager(&rpm);
-
-    // Two overlay passes in one frame — the history sample is the
-    // frame's TOTAL overlay cost (1.5 + 0.5 = 2.0ms, 4 + 2 = 6 draws).
-    render_synthetic_frame(rpm, {
-        {RenderPassType::overlay, 1.5f, 4},
-        {RenderPassType::overlay, 0.5f, 2},
-    });
-    REQUIRE(overlay.capture_pass_frame());
-
-    auto attrib = overlay.pass_attribution();
-    const auto& ov = attrib[static_cast<int>(RenderPassType::overlay)];
-    REQUIRE(ov.samples == 1);
-    REQUIRE_THAT(ov.last_cpu_ms, Catch::Matchers::WithinAbs(2.0, 0.001));
-    REQUIRE(ov.last_draw_calls == 6);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: history ring caps at kPassHistoryFrames",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-
-    RenderPassManager rpm;
-    overlay.set_render_pass_manager(&rpm);
-
-    // Push more frames than the ring holds; only the last N count
-    // toward per-pass detail, but the lifetime counter keeps growing.
-    const std::size_t extra = 25;
-    const std::size_t total = InspectorOverlay::kPassHistoryFrames + extra;
-    for (std::size_t i = 0; i < total; ++i) {
-        render_synthetic_frame(rpm, {{RenderPassType::content,
-                                      static_cast<float>(i % 7), 1}});
-        REQUIRE(overlay.capture_pass_frame());
-    }
-    REQUIRE(overlay.pass_frames_captured() == total);
-
-    auto attrib = overlay.pass_attribution();
-    const auto& content = attrib[static_cast<int>(RenderPassType::content)];
-    REQUIRE(content.samples == InspectorOverlay::kPassHistoryFrames);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: budget overruns are counted",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-
-    RenderPassManager rpm;
-    rpm.set_budget(5.0f);  // tight 5ms budget.
-    overlay.set_render_pass_manager(&rpm);
-
-    // Frame 1: under budget (3ms).
-    render_synthetic_frame(rpm, {{RenderPassType::content, 3.0f, 4}});
-    REQUIRE(overlay.capture_pass_frame());
-    REQUIRE(overlay.budget_overrun_count() == 0);
-
-    // Frame 2: over budget (9ms).
-    render_synthetic_frame(rpm, {{RenderPassType::content, 9.0f, 4}});
-    REQUIRE(overlay.capture_pass_frame());
-    REQUIRE(overlay.budget_overrun_count() == 1);
-
-    // Frame 3: over budget again (12ms).
-    render_synthetic_frame(rpm, {{RenderPassType::content, 12.0f, 4}});
-    REQUIRE(overlay.capture_pass_frame());
-    REQUIRE(overlay.budget_overrun_count() == 2);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: P key toggles the attribution viewer",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    REQUIRE_FALSE(overlay.pass_viewer_enabled());
-
-    KeyEvent p;
-    p.key = KeyCode::p;
-    p.is_down = true;
-    p.modifiers = 0;
-    REQUIRE(overlay.handle_key_event(p));
-    REQUIRE(overlay.pass_viewer_enabled());
-
-    REQUIRE(overlay.handle_key_event(p));
-    REQUIRE_FALSE(overlay.pass_viewer_enabled());
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: P key ignored when inspector inactive",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-    // Not active — the P hotkey must not fire so it can't collide with
-    // a plain text field in the host UI.
-
-    KeyEvent p;
-    p.key = KeyCode::p;
-    p.is_down = true;
-    p.modifiers = 0;
-    REQUIRE_FALSE(overlay.handle_key_event(p));
-    REQUIRE_FALSE(overlay.pass_viewer_enabled());
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: viewer renders per-pass rows",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_id("root");
-    // Tall enough that the panel's lower section fits both pass rows
-    // (heading + summary + two ~56px rows). The lower section gets
-    // roughly half the window height minus the stats bar.
-    root.set_bounds({0, 0, 500, 720});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    RenderPassManager rpm;
-    overlay.set_render_pass_manager(&rpm);
-
-    // Drive a few frames with two distinct pass types.
-    for (int i = 0; i < 3; ++i) {
-        render_synthetic_frame(rpm, {
-            {RenderPassType::background, 1.0f, 2},
-            {RenderPassType::content,    5.0f + static_cast<float>(i), 14},
-        });
-        // paint() captures the frame; toggle viewer on first.
-        if (i == 0) overlay.set_pass_viewer_enabled(true);
-        pulp::canvas::RecordingCanvas warmup;
-        overlay.paint(warmup);
-    }
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-    REQUIRE(canvas.command_count() > 0);
-
-    // The viewer heading and both rendered pass names must appear.
-    REQUIRE(count_text_containing(canvas, "Render Passes") >= 1);
-    REQUIRE(count_text_containing(canvas, "background") >= 1);
-    REQUIRE(count_text_containing(canvas, "content") >= 1);
-    // Honesty note about CPU vs GPU timing is surfaced.
-    REQUIRE(count_text_containing(canvas, "Phase 6.5") >= 1);
-    // "effects" never rendered — must NOT show a row for it.
-    REQUIRE(count_text_containing(canvas, "effects") == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: viewer reports missing RPM gracefully",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 500, 320});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    overlay.set_pass_viewer_enabled(true);
-    // No RenderPassManager attached.
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-    REQUIRE(count_text_containing(canvas, "No RenderPassManager") >= 1);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.1: paint() drives capture automatically",
-          "[inspect][overlay][phase6]") {
-    View root;
-    root.set_bounds({0, 0, 500, 320});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    RenderPassManager rpm;
-    overlay.set_render_pass_manager(&rpm);
-
-    render_synthetic_frame(rpm, {{RenderPassType::content, 4.0f, 8}});
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);  // should capture the frame internally.
-    REQUIRE(overlay.pass_frames_captured() == 1);
-
-    // A second paint of the SAME frame does not re-capture.
-    overlay.paint(canvas);
-    REQUIRE(overlay.pass_frames_captured() == 1);
-
-    // A new frame, then paint, captures again.
-    render_synthetic_frame(rpm, {{RenderPassType::content, 6.0f, 9}});
-    overlay.paint(canvas);
-    REQUIRE(overlay.pass_frames_captured() == 2);
-}
-
-// ── Phase 5.1: source-jump (View provenance + overlay J hotkey) ────────────
-//
-// planning/2026-05-19-inspector-phase5-source-jump-spike.md § Phase 5.1.
-
-#include <pulp/inspect/source_jump.hpp>
-#include <choc/text/choc_JSON.h>
-
-#include <cstdlib>
-
-namespace {
-
-// Scoped guard that sets PULP_INSPECTOR_NO_LAUNCH for the duration of a
-// test, restoring the prior value on destruction. The overlay's J
-// hotkey resolves with dry_run=false and would otherwise spawn a real
-// editor (and pop the macOS open-confirmation dialog). The CTest target
-// sets this env var globally, but a test that exercises the real-launch
-// path sets it in-process too so it is safe when the binary is run
-// directly. With the guard in place, launch_editor_url() never spawns a
-// process and reports launched=false.
-struct ScopedNoLaunch {
-    ScopedNoLaunch() {
-        if (const char* prev = std::getenv("PULP_INSPECTOR_NO_LAUNCH")) {
-            prev_ = std::string(prev);
-            had_prev_ = true;
-        }
-#if defined(_WIN32)
-        _putenv_s("PULP_INSPECTOR_NO_LAUNCH", "1");
-#else
-        ::setenv("PULP_INSPECTOR_NO_LAUNCH", "1", /*overwrite=*/1);
-#endif
-    }
-    ~ScopedNoLaunch() {
-#if defined(_WIN32)
-        _putenv_s("PULP_INSPECTOR_NO_LAUNCH", had_prev_ ? prev_.c_str() : "");
-#else
-        if (had_prev_) ::setenv("PULP_INSPECTOR_NO_LAUNCH", prev_.c_str(), 1);
-        else ::unsetenv("PULP_INSPECTOR_NO_LAUNCH");
-#endif
-    }
-    std::string prev_;
-    bool had_prev_ = false;
-};
-
-} // namespace
-
-TEST_CASE("View::source_loc round-trips a file:line:col record",
-          "[inspect][source-jump]") {
-    View v;
-    REQUIRE_FALSE(v.has_source_loc());
-    REQUIRE_FALSE(v.source_loc().valid());
-
-    v.set_source_loc({"src/Synth.jsx", 42, 7});
-    REQUIRE(v.has_source_loc());
-    REQUIRE(v.source_loc().valid());
-    REQUIRE(v.source_loc().file == "src/Synth.jsx");
-    REQUIRE(v.source_loc().line == 42);
-    REQUIRE(v.source_loc().col == 7);
-
-    v.clear_source_loc();
-    REQUIRE_FALSE(v.has_source_loc());
-}
-
-TEST_CASE("ViewInspector::find_by_anchor locates a view by its anchor id",
-          "[inspect][source-jump]") {
-    View root;
-    auto a = std::make_unique<View>();
-    a->set_anchor_id("figma:a");
-    auto* a_ptr = a.get();
-    auto b = std::make_unique<View>();
-    b->set_anchor_id("figma:b");
-    auto* b_ptr = b.get();
-    root.add_child(std::move(a));
-    root.add_child(std::move(b));
-
-    REQUIRE(ViewInspector::find_by_anchor(root, "figma:a") == a_ptr);
-    REQUIRE(ViewInspector::find_by_anchor(root, "figma:b") == b_ptr);
-    REQUIRE(ViewInspector::find_by_anchor(root, "missing") == nullptr);
-    // An empty anchor never matches a (default-empty-anchor) view.
-    REQUIRE(ViewInspector::find_by_anchor(root, "") == nullptr);
-}
-
-TEST_CASE("InspectorOverlay: J jumps to source for a view with provenance",
-          "[inspect][overlay][source-jump]") {
-    // The J hotkey resolves with dry_run=false — the real-launch path.
-    // Without the no-launch guard this would spawn `open vscode://...`
-    // and pop the macOS open-confirmation dialog. The guard makes
-    // launch_editor_url() a no-op so the test verifies the *constructed*
-    // URL, never an actual editor launch.
-    ScopedNoLaunch no_launch;
-
-    View root;
-    root.set_bounds({0, 0, 500, 300});
-    auto child = std::make_unique<View>();
-    child->set_id("knob");
-    child->set_bounds({10, 10, 80, 40});
-    child->set_source_loc({"src/Panel.jsx", 24, 3});
-    auto* child_ptr = child.get();
-    root.add_child(std::move(child));
-
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    // Select the child via a click.
-    MouseEvent click;
-    click.position = {20, 20};
-    click.is_down = true;
-    REQUIRE(overlay.handle_mouse_event(click));
-    REQUIRE(overlay.selected_view() == child_ptr);
-
-    // Dry-run jump — resolves the URL but never spawns the editor.
-    auto result = overlay.jump_to_selection_source(/*dry_run=*/true);
-    REQUIRE(result.ok);
-    REQUIRE(result.url == "vscode://file/src/Panel.jsx:24");
-    REQUIRE_FALSE(result.launched);
-
-    // Real-launch path (dry_run=false), as the J hotkey runs it: the URL
-    // still resolves, but the no-launch guard suppresses the spawn so
-    // `launched` stays false. This is the regression assertion for the
-    // unexpected-editor-launch bug.
-    auto live = overlay.jump_to_selection_source(/*dry_run=*/false);
-    REQUIRE(live.ok);
-    REQUIRE(live.url == "vscode://file/src/Panel.jsx:24");
-    REQUIRE_FALSE(live.launched);
-
-    // launch_editor_url() itself is a no-op under the guard.
-    REQUIRE_FALSE(launch_editor_url("vscode://file/src/Panel.jsx:24"));
-
-    // The J hotkey is consumed while the inspector is active. The
-    // handler runs the dry_run=false path; the guard keeps it from
-    // spawning a real editor process.
-    KeyEvent jk;
-    jk.key = KeyCode::j;
-    jk.modifiers = 0;
-    jk.is_down = true;
-    REQUIRE(overlay.handle_key_event(jk));
-}
-
-TEST_CASE("InspectorOverlay source jump helper defaults to dry-run",
-          "[inspect][overlay][source-jump][issue-2515]") {
-    ScopedEnv headless("PULP_HEADLESS");
-    headless.set("1");
-
-    View root;
-    root.set_bounds({0, 0, 500, 300});
-    auto child = std::make_unique<View>();
-    child->set_bounds({10, 10, 80, 40});
-    child->set_source_loc({"src/Panel.jsx", 24, 3});
-    root.add_child(std::move(child));
-
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    MouseEvent click;
-    click.position = {20, 20};
-    click.is_down = true;
-    REQUIRE(overlay.handle_mouse_event(click));
-    REQUIRE(overlay.selected_view() != nullptr);
-
-    auto result = overlay.jump_to_selection_source();
-    REQUIRE(result.ok);
-    REQUIRE(result.url == "vscode://file/src/Panel.jsx:24");
-    REQUIRE_FALSE(result.launched);
-    REQUIRE(result.error.empty());
-}
-
-TEST_CASE("InspectorOverlay: J is a graceful no-op without a selection",
-          "[inspect][overlay][source-jump]") {
-    // No selection means jump_to_source resolves ok==false and never
-    // reaches launch_editor_url(); the guard is belt-and-suspenders so
-    // the key event below stays inert when the binary is run directly.
-    ScopedNoLaunch no_launch;
-
-    View root;
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    REQUIRE(overlay.selected_view() == nullptr);
-
-    auto result = overlay.jump_to_selection_source(/*dry_run=*/true);
-    REQUIRE_FALSE(result.ok);
-    REQUIRE(result.url.empty());
-
-    // The key is still consumed (does not fall through to the view tree),
-    // but no jump happens.
-    KeyEvent jk;
-    jk.key = KeyCode::j;
-    jk.modifiers = 0;
-    jk.is_down = true;
-    REQUIRE(overlay.handle_key_event(jk));
-}
-
-TEST_CASE("InspectorOverlay: J is a graceful no-op for a non-imported view",
-          "[inspect][overlay][source-jump]") {
-    View root;
-    root.set_bounds({0, 0, 500, 300});
-    auto child = std::make_unique<View>();
-    child->set_bounds({10, 10, 80, 40});
-    // deliberately NO set_source_loc — a user-authored, non-JSX view.
-    root.add_child(std::move(child));
-
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    MouseEvent click;
-    click.position = {20, 20};
-    click.is_down = true;
-    overlay.handle_mouse_event(click);
-    REQUIRE(overlay.selected_view() != nullptr);
-
-    auto result = overlay.jump_to_selection_source(/*dry_run=*/true);
-    REQUIRE_FALSE(result.ok);
-    REQUIRE(result.error.find("no source location") != std::string::npos);
-}
-
-TEST_CASE("InspectorOverlay: J does nothing when the inspector is inactive",
-          "[inspect][overlay][source-jump]") {
-    View root;
-    InspectorOverlay overlay(root);
-    // inspector NOT active
-    KeyEvent jk;
-    jk.key = KeyCode::j;
-    jk.modifiers = 0;
-    jk.is_down = true;
-    REQUIRE_FALSE(overlay.handle_key_event(jk));
-}
-
-TEST_CASE("InspectorOverlay: config swap changes the source-jump template",
-          "[inspect][overlay][source-jump]") {
-    View root;
-    root.set_bounds({0, 0, 500, 300});
-    auto child = std::make_unique<View>();
-    child->set_bounds({10, 10, 80, 40});
-    child->set_source_loc({"a/B.tsx", 11, 2});
-    root.add_child(std::move(child));
-
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    MouseEvent click;
-    click.position = {20, 20};
-    click.is_down = true;
-    overlay.handle_mouse_event(click);
-
-    InspectorConfig cfg;
-    cfg.editor_url_template = "zed://file/{path}:{line}:{col}";
-    overlay.set_config(cfg);
-
-    auto result = overlay.jump_to_selection_source(/*dry_run=*/true);
-    REQUIRE(result.ok);
-    REQUIRE(result.url == "zed://file/a/B.tsx:11:2");
-}
-
-TEST_CASE("DomainHandler propagates config to the attached overlay",
-          "[inspect][source-jump][domain]") {
-    View root;
-    root.set_bounds({0, 0, 500, 300});
-    auto child = std::make_unique<View>();
-    child->set_anchor_id("anchor-1");
-    child->set_bounds({10, 10, 80, 40});
-    child->set_source_loc({"x/Y.jsx", 5, 0});
-    root.add_child(std::move(child));
-
-    InspectorOverlay overlay(root);
-    DomainHandler handler;
-    handler.set_root_view(&root);
-    handler.set_overlay(&overlay);
-
-    InspectorConfig cfg;
-    cfg.editor_url_template = "cursor://file/{path}:{line}";
-    handler.set_config(cfg);
-
-    // The overlay's config now matches the handler's.
-    REQUIRE(overlay.config().editor_url_template
-            == "cursor://file/{path}:{line}");
-
-    // And Inspector.jumpToSource via the anchor resolves with it.
-    auto resp = handler.handle(make_request(
-        1, methods::kInspectorJumpToSource,
-        R"({"anchorId":"anchor-1","dryRun":true})"));
-    REQUIRE_FALSE(resp.is_error);
-    auto obj = choc::json::parse(resp.params_json);
-    REQUIRE(obj["ok"].getBool());
-    REQUIRE(std::string(obj["url"].getString())
-            == "cursor://file/x/Y.jsx:5");
-}
-
-// ── Phase 2 — drift drawer ──────────────────────────────────────────────
-//
-// The drift drawer is a collapsible warning panel inside the inspector
-// overlay that lists tweaks whose anchor_id no longer resolves to a live
-// view. It is populated by refresh_drift(), which walks the live view
-// tree's anchor set and diffs it against the attached TweakStore.
-// Spec: planning/2026-05-18-inspector-direct-manipulation-roadmap.md
-// (Phase 2).
-namespace {
-
-// True if any fill_text command's text contains `needle`.
-bool canvas_has_text(const pulp::canvas::RecordingCanvas& canvas,
-                     std::string_view needle) {
-    for (const auto& cmd : canvas.commands()) {
-        if (cmd.type == pulp::canvas::DrawCommand::Type::fill_text &&
-            cmd.text.find(needle) != std::string::npos) {
-            return true;
-        }
-    }
-    return false;
-}
-
-}  // namespace
-
-TEST_CASE("InspectorOverlay: refresh_drift finds tweaks with no live anchor",
-          "[inspect][overlay][drift][phase2]") {
-    View root;
-    root.set_bounds({0, 0, 500, 400});
-
-    auto child = std::make_unique<View>();
-    child->set_anchor_id("anchor-still-here");
-    root.add_child(std::move(child));
-
-    TweakStore store;
-    // One tweak whose anchor is live, one whose anchor is gone.
-    store.apply_tweak("anchor-still-here", "layout.padding",
-                      choc::value::createInt32(8));
-    store.apply_tweak("anchor-removed-by-reimport", "paint.color",
-                      choc::value::createString("#ff0000"));
-
-    InspectorOverlay overlay(root);
-    overlay.set_tweak_store(&store);
-    overlay.refresh_drift();
-
-    REQUIRE(overlay.drift_count() == 1);
-    REQUIRE(overlay.drifted().front().anchor_id ==
-            "anchor-removed-by-reimport");
-    REQUIRE(overlay.drifted().front().reason ==
-            TweakStore::DriftReason::anchor_not_found);
-    // First drift detection auto-expands the drawer so it is not silent.
-    REQUIRE(overlay.drift_drawer_open());
-}
-
-TEST_CASE("InspectorOverlay: no drift means no drawer and an empty list",
-          "[inspect][overlay][drift][phase2]") {
-    View root;
-    root.set_bounds({0, 0, 500, 400});
-    auto child = std::make_unique<View>();
-    child->set_anchor_id("live-anchor");
-    root.add_child(std::move(child));
-
-    TweakStore store;
-    store.apply_tweak("live-anchor", "layout.width",
-                      choc::value::createInt32(120));
-
-    InspectorOverlay overlay(root);
-    overlay.set_tweak_store(&store);
-    overlay.refresh_drift();
-
-    REQUIRE(overlay.drift_count() == 0);
-    REQUIRE(overlay.drifted().empty());
-
-    // The drawer paints nothing on the happy path.
-    overlay.set_active(true);
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-    REQUIRE_FALSE(canvas_has_text(canvas, "Drift"));
-}
-
-TEST_CASE("InspectorOverlay: refresh_drift with no TweakStore is a safe no-op",
-          "[inspect][overlay][drift][phase2]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-    overlay.refresh_drift();  // no store wired
-    REQUIRE(overlay.drift_count() == 0);
-    REQUIRE(overlay.drifted().empty());
-}
-
-TEST_CASE("InspectorOverlay: drift drawer renders header and orphan rows",
-          "[inspect][overlay][drift][phase2]") {
-    View root;
-    root.set_bounds({0, 0, 600, 500});
-    auto child = std::make_unique<View>();
-    child->set_anchor_id("kept");
-    root.add_child(std::move(child));
-
-    TweakStore store;
-    store.apply_tweak("kept", "layout.width",
-                      choc::value::createInt32(64));
-    store.apply_tweak("gone-anchor", "layout.margin",
-                      choc::value::createInt32(10));
-
-    InspectorOverlay overlay(root);
-    overlay.set_tweak_store(&store);
-    overlay.set_active(true);
-
-    // First paint auto-refreshes drift and auto-expands the drawer.
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-
-    REQUIRE(overlay.drift_count() == 1);
-    REQUIRE(overlay.drift_drawer_open());
-    // Header text + the orphaned anchor + its reason tag are painted.
-    REQUIRE(canvas_has_text(canvas, "Drift"));
-    REQUIRE(canvas_has_text(canvas, "gone-anchor"));
-    REQUIRE(canvas_has_text(canvas, "anchor-not-found"));
-    REQUIRE(canvas_has_text(canvas, "layout.margin"));
-}
-
-TEST_CASE("InspectorOverlay: clicking the drift header toggles the drawer",
-          "[inspect][overlay][drift][phase2]") {
-    View root;
-    root.set_bounds({0, 0, 600, 500});
-    auto child = std::make_unique<View>();
-    child->set_anchor_id("present");
-    root.add_child(std::move(child));
-
-    TweakStore store;
-    store.apply_tweak("present", "layout.width",
-                      choc::value::createInt32(80));
-    store.apply_tweak("missing", "paint.opacity",
-                      choc::value::createFloat32(0.4f));
-
-    InspectorOverlay overlay(root);
-    overlay.set_tweak_store(&store);
-    overlay.set_active(true);
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);  // auto-expands; populates the header hit-rect
-    REQUIRE(overlay.drift_drawer_open());
-
-    // The drawer header sits just below the tree section (root_h * 0.5
-    // = 250) on the panel side (panel_x = 600 - 300 = 300). A click in
-    // that band collapses the drawer.
-    MouseEvent click;
-    click.position = {360, 258};
-    click.is_down = true;
-    REQUIRE(overlay.handle_mouse_event(click));
-    REQUIRE_FALSE(overlay.drift_drawer_open());
-
-    // Re-paint so the (now-collapsed) header hit-rect is refreshed,
-    // then click again to re-expand.
-    canvas.clear();
-    overlay.paint(canvas);
-    MouseEvent click2;
-    click2.position = {360, 258};
-    click2.is_down = true;
-    REQUIRE(overlay.handle_mouse_event(click2));
-    REQUIRE(overlay.drift_drawer_open());
-}
-
-TEST_CASE("InspectorOverlay: collapsed drawer hides rows but keeps the header",
-          "[inspect][overlay][drift][phase2]") {
-    View root;
-    root.set_bounds({0, 0, 600, 500});
-
-    TweakStore store;
-    store.apply_tweak("orphan-a", "layout.width",
-                      choc::value::createInt32(50));
-
-    InspectorOverlay overlay(root);
-    overlay.set_tweak_store(&store);
-    overlay.set_active(true);
-    overlay.refresh_drift();
-    overlay.set_drift_drawer_open(false);
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-    // Header still shows the drift count; the orphan row does not.
-    REQUIRE(canvas_has_text(canvas, "Drift"));
-    REQUIRE_FALSE(canvas_has_text(canvas, "layout.width"));
-}
-
-// ── Phase 5.2 — reconciliation tab ────────────────────────────────────────
-//
-// The reconciliation tab (R-key) classifies every stored tweak into
-// locked-to-source / drifted / unresolvable and renders a read-only
-// report. It builds on the Phase 4a lock-to-source state (the
-// TweakStore lock set) and the live view tree's anchor set. These
-// tests drive it headless: build a scene, seed tweaks, toggle the tab
-// with R, paint into a RecordingCanvas, and assert the classification.
-
-namespace {
-
-// A scene with two live anchored views plus one tweak whose anchor is
-// NOT in the tree (orphaned). Locking one of the live anchors lets a
-// test exercise all three reconciliation states at once.
-struct ReconcileScene {
-    View root;
-    TweakStore store;
-    InspectorOverlay overlay{root};
-
-    ReconcileScene() {
-        root.set_bounds({0, 0, 600, 600});
-
-        auto a = std::make_unique<View>();
-        a->set_anchor_id("figma:5:a");
-        a->set_bounds({10, 10, 80, 40});
-        root.add_child(std::move(a));
-
-        auto b = std::make_unique<View>();
-        b->set_anchor_id("figma:5:b");
-        b->set_bounds({10, 60, 80, 40});
-        root.add_child(std::move(b));
-
-        overlay.set_active(true);
-        overlay.set_tweak_store(&store);
-
-        // figma:5:a — locked → "locked-to-source" once locked below.
-        store.apply_tweak("figma:5:a", "layout.padding",
-                          choc::value::createInt32(12), "drag");
-        // figma:5:b — resolves but unlocked → "drifted".
-        store.apply_tweak("figma:5:b", "layout.gap",
-                          choc::value::createInt32(4), "drag");
-        // figma:5:gone — no live view carries it → "unresolvable".
-        store.apply_tweak("figma:5:gone", "paint.backgroundColor",
-                          choc::value::createString("#abcdef"), "picker");
-    }
-};
-
-}  // namespace
-
-TEST_CASE("InspectorOverlay Phase 5.2: R toggles the reconciliation tab",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    View root;
-    root.set_bounds({0, 0, 600, 600});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    REQUIRE_FALSE(overlay.reconcile_tab_visible());
-
-    KeyEvent r;
-    r.key = KeyCode::r;
-    r.is_down = true;
-    REQUIRE(overlay.handle_key_event(r));
-    REQUIRE(overlay.reconcile_tab_visible());
-
-    REQUIRE(overlay.handle_key_event(r));
-    REQUIRE_FALSE(overlay.reconcile_tab_visible());
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: R does nothing when inspector inactive",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    View root;
-    InspectorOverlay overlay(root);
-    // Inspector not active — R must not flip the tab.
-    REQUIRE_FALSE(overlay.handle_key_event(make_key(KeyCode::r)));
-    REQUIRE_FALSE(overlay.reconcile_tab_visible());
-
-    // A modifier-laden R is reserved for chord shortcuts — ignored.
-    overlay.set_active(true);
-    overlay.handle_key_event(make_key(KeyCode::r, true, kModCmd));
-    REQUIRE_FALSE(overlay.reconcile_tab_visible());
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: reconcile_report classifies all states",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    ReconcileScene scene;
-    // Lock figma:5:a — promotes its tweak to "locked-to-source".
-    scene.store.set_locked("figma:5:a", true);
-
-    auto report = scene.overlay.reconcile_report();
-    REQUIRE(report.total() == 3);
-    REQUIRE(report.locked_count == 1);
-    REQUIRE(report.drifted_count == 1);
-    REQUIRE(report.unresolvable_count == 1);
-
-    // Verify the per-row classification by anchor.
-    auto status_for = [&](std::string_view anchor)
-        -> InspectorOverlay::ReconcileStatus {
-        for (const auto& row : report.rows)
-            if (row.anchor_id == anchor) return row.status;
-        FAIL("anchor not found in report");
-        return InspectorOverlay::ReconcileStatus::unresolvable;
-    };
-    REQUIRE(status_for("figma:5:a") ==
-            InspectorOverlay::ReconcileStatus::locked_to_source);
-    REQUIRE(status_for("figma:5:b") ==
-            InspectorOverlay::ReconcileStatus::drifted);
-    REQUIRE(status_for("figma:5:gone") ==
-            InspectorOverlay::ReconcileStatus::unresolvable);
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: unlocked live tweak reads as drifted",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    // Without an explicit lock, a tweak whose anchor resolves to a
-    // live view is "drifted" — it lives only in the runtime layer and
-    // would not survive a re-import.
-    ReconcileScene scene;  // no set_locked() call
-    auto report = scene.overlay.reconcile_report();
-    REQUIRE(report.locked_count == 0);
-    // figma:5:a + figma:5:b both resolve but are unlocked → drifted.
-    REQUIRE(report.drifted_count == 2);
-    REQUIRE(report.unresolvable_count == 1);
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: tab lays out a row per tweak",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    ReconcileScene scene;
-    scene.store.set_locked("figma:5:a", true);
-    scene.overlay.toggle_reconcile_tab();
-    REQUIRE(scene.overlay.reconcile_tab_visible());
-
-    pulp::canvas::RecordingCanvas canvas;
-    scene.overlay.paint(canvas);
-    REQUIRE(canvas.command_count() > 0);
-
-    // Three tweaks → three classified rows laid out.
-    REQUIRE(scene.overlay.reconcile_row_count() == 3);
-    // The tab heading + status tags render as text.
-    REQUIRE(canvas_has_text(canvas, "Reconcile"));
-    REQUIRE(canvas_has_text(canvas, "lock"));
-    REQUIRE(canvas_has_text(canvas, "drift"));
-    REQUIRE(canvas_has_text(canvas, "orphan"));
-
-    // Hiding the tab clears the laid-out rows.
-    scene.overlay.toggle_reconcile_tab();
-    pulp::canvas::RecordingCanvas hidden;
-    scene.overlay.paint(hidden);
-    REQUIRE(scene.overlay.reconcile_row_count() == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: tab renders cleanly with no tweaks",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    View root;
-    root.set_bounds({0, 0, 600, 600});
-    TweakStore store;  // empty
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    overlay.set_tweak_store(&store);
-    overlay.toggle_reconcile_tab();
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);  // must not crash
-    REQUIRE(canvas.command_count() > 0);
-    REQUIRE(overlay.reconcile_row_count() == 0);
-    REQUIRE(canvas_has_text(canvas, "No tweaks to reconcile"));
-
-    auto report = overlay.reconcile_report();
-    REQUIRE(report.total() == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: reconcile_report empty without a store",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    // The inspector can run with no TweakStore wired (hand-authored
-    // UIs). The reconciliation tab must degrade gracefully — empty
-    // report, no crash, an explanatory line in the panel.
-    View root;
-    root.set_bounds({0, 0, 600, 600});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    overlay.toggle_reconcile_tab();
-    // intentionally no set_tweak_store()
-
-    auto report = overlay.reconcile_report();
-    REQUIRE(report.total() == 0);
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-    REQUIRE(canvas.command_count() > 0);
-    REQUIRE(overlay.reconcile_row_count() == 0);
-    REQUIRE(canvas_has_text(canvas, "No tweak store attached"));
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: orphaned tweak never crashes the tab",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    // A tweak whose anchor has no live view (e.g. after a destructive
-    // re-import) must classify as unresolvable — not crash, not guess.
-    View root;
-    root.set_bounds({0, 0, 600, 600});
-    TweakStore store;
-    store.apply_tweak("ghost-anchor", "layout.width",
-                      choc::value::createInt32(99), "drag");
-
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    overlay.set_tweak_store(&store);
-    overlay.toggle_reconcile_tab();
-
-    auto report = overlay.reconcile_report();
-    REQUIRE(report.total() == 1);
-    REQUIRE(report.unresolvable_count == 1);
-    REQUIRE(report.rows.front().status ==
-            InspectorOverlay::ReconcileStatus::unresolvable);
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);  // must not crash
-    REQUIRE(overlay.reconcile_row_count() == 1);
-    REQUIRE(canvas_has_text(canvas, "orphan"));
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: reconcile_status_str maps every state",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    using RS = InspectorOverlay::ReconcileStatus;
-    REQUIRE(std::string(InspectorOverlay::reconcile_status_str(
-                RS::locked_to_source)) == "locked-to-source");
-    REQUIRE(std::string(InspectorOverlay::reconcile_status_str(
-                RS::drifted)) == "drifted");
-    REQUIRE(std::string(InspectorOverlay::reconcile_status_str(
-                RS::unresolvable)) == "unresolvable");
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: dismissing the inspector clears tab rows",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    ReconcileScene scene;
-    scene.overlay.toggle_reconcile_tab();
-
-    pulp::canvas::RecordingCanvas canvas;
-    scene.overlay.paint(canvas);
-    REQUIRE(scene.overlay.reconcile_row_count() == 3);
-
-    // Closing the inspector drops the laid-out rows; the R-toggle
-    // state itself survives so re-opening restores the tab.
-    scene.overlay.set_active(false);
-    REQUIRE(scene.overlay.reconcile_row_count() == 0);
-    REQUIRE(scene.overlay.reconcile_tab_visible());
-}
-
-TEST_CASE("InspectorOverlay Phase 5.2: locking a tweak moves it out of drift",
-          "[inspect][overlay][reconcile][phase5.2]") {
-    // Locking an anchor is the explicit "promote to source" action.
-    // The reconciliation tab must reflect that transition: a drifted
-    // tweak becomes locked-to-source the moment its anchor is locked.
-    ReconcileScene scene;
-    auto before = scene.overlay.reconcile_report();
-    REQUIRE(before.locked_count == 0);
-    REQUIRE(before.drifted_count == 2);
-
-    scene.store.set_locked("figma:5:b", true);
-    auto after = scene.overlay.reconcile_report();
-    REQUIRE(after.locked_count == 1);
-    REQUIRE(after.drifted_count == 1);
-    REQUIRE(after.unresolvable_count == 1);  // figma:5:gone unchanged
-}
-
-// ── Phase 6.2 — texture atlas viewer ────────────────────────────────────────
-//
-// Spec: planning/2026-05-19-inspector-phase6-gpu-perf-spike.md § Phase 6.2.
-// The atlas viewer is a read-only GPU-perf observability tab over the
-// render layer's AtlasInventory. These headless tests drive it through
-// the RecordingCanvas — the tab toggles via the `A` hotkey, renders one
-// row per registered atlas, and degrades gracefully when no inventory is
-// attached (GPU-off / headless builds). It invents no data: every figure
-// comes from the Pulp-owned atlas classes' introspection accessors.
-
-namespace {
-
-using pulp::render::AtlasInfo;
-using pulp::render::AtlasInventory;
-using pulp::render::AtlasKind;
-using pulp::render::GlyphAtlas;
-using pulp::render::ImageAtlas;
-
-// Build a small inventory snapshotting two real, partially-packed atlases
-// so the viewer has honest dimensions + occupancy to render.
-AtlasInventory make_atlas_inventory() {
-    static ImageAtlas images(256);
-    static GlyphAtlas glyphs(512);
-    pulp::render::AtlasPacker::Region r{};
-    images.allocate(1, 128, 256, r);   // 50% of the image atlas.
-    glyphs.allocate(2, 512, 256, r);   // 50% of the glyph atlas.
-
-    AtlasInventory inv;
-    inv.add_atlas(images, AtlasKind::image, "images");
-    inv.add_atlas(glyphs, AtlasKind::glyph);
-    return inv;
-}
-
-} // namespace
-
-TEST_CASE("InspectorOverlay Phase 6.2: A key toggles the atlas viewer",
-          "[inspect][overlay][atlas][phase6.2]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    REQUIRE_FALSE(overlay.atlas_viewer_visible());
-
-    KeyEvent a;
-    a.key = KeyCode::a;
-    a.is_down = true;
-    a.modifiers = 0;
-    REQUIRE(overlay.handle_key_event(a));
-    REQUIRE(overlay.atlas_viewer_visible());
-
-    REQUIRE(overlay.handle_key_event(a));
-    REQUIRE_FALSE(overlay.atlas_viewer_visible());
-}
-
-TEST_CASE("InspectorOverlay Phase 6.2: A key ignored when inspector inactive",
-          "[inspect][overlay][atlas][phase6.2]") {
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    InspectorOverlay overlay(root);
-    // Not active — the A hotkey must not fire so it can't collide with
-    // a plain text field in the host UI.
-
-    KeyEvent a;
-    a.key = KeyCode::a;
-    a.is_down = true;
-    a.modifiers = 0;
-    REQUIRE_FALSE(overlay.handle_key_event(a));
-    REQUIRE_FALSE(overlay.atlas_viewer_visible());
-}
-
-TEST_CASE("InspectorOverlay Phase 6.2: viewer renders a row per atlas",
-          "[inspect][overlay][atlas][phase6.2]") {
-    View root;
-    root.set_id("root");
-    // Tall enough that the panel's lower section fits both atlas rows.
-    root.set_bounds({0, 0, 500, 720});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    AtlasInventory inv = make_atlas_inventory();
-    overlay.set_atlas_inventory(&inv);
-    overlay.set_atlas_viewer_visible(true);
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-    REQUIRE(canvas.command_count() > 0);
-
-    // The viewer heading and both atlas labels must appear.
-    REQUIRE(count_text_containing(canvas, "Texture Atlases") >= 1);
-    REQUIRE(count_text_containing(canvas, "images") >= 1);
-    REQUIRE(count_text_containing(canvas, "glyph") >= 1);
-    // Each atlas printed an occupancy percentage — both atlases are at
-    // 50% so a "50%" readout is present.
-    REQUIRE(count_text_containing(canvas, "50%") >= 1);
-    // Two atlases were registered → two laid-out rows.
-    REQUIRE(overlay.atlas_row_count() == 2);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.2: viewer degrades gracefully with no inventory",
-          "[inspect][overlay][atlas][phase6.2]") {
-    View root;
-    root.set_bounds({0, 0, 500, 320});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    overlay.set_atlas_viewer_visible(true);
-    // No AtlasInventory attached — the GPU-off / headless path.
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);  // must not crash.
-    REQUIRE(count_text_containing(canvas, "GPU atlas unavailable") >= 1);
-    REQUIRE(overlay.atlas_row_count() == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.2: empty inventory shows the unavailable state",
-          "[inspect][overlay][atlas][phase6.2]") {
-    View root;
-    root.set_bounds({0, 0, 500, 320});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-    overlay.set_atlas_viewer_visible(true);
-
-    AtlasInventory empty;  // attached but holds no atlases.
-    overlay.set_atlas_inventory(&empty);
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-    REQUIRE(count_text_containing(canvas, "GPU atlas unavailable") >= 1);
-    REQUIRE(overlay.atlas_row_count() == 0);
-}
-
-TEST_CASE("InspectorOverlay Phase 6.2: dismissing the inspector clears atlas rows",
-          "[inspect][overlay][atlas][phase6.2]") {
-    View root;
-    root.set_bounds({0, 0, 500, 720});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    AtlasInventory inv = make_atlas_inventory();
-    overlay.set_atlas_inventory(&inv);
-    overlay.set_atlas_viewer_visible(true);
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-    REQUIRE(overlay.atlas_row_count() == 2);
-
-    // Closing the inspector resets the laid-out row count but leaves the
-    // A-key toggle intact so re-opening restores the same tab.
-    overlay.set_active(false);
-    REQUIRE(overlay.atlas_row_count() == 0);
-    REQUIRE(overlay.atlas_viewer_visible());
-}
-
-TEST_CASE("InspectorOverlay Phase 6.2: pass viewer takes precedence over atlas tab",
-          "[inspect][overlay][atlas][phase6.2]") {
-    View root;
-    root.set_bounds({0, 0, 500, 720});
-    InspectorOverlay overlay(root);
-    overlay.set_active(true);
-
-    AtlasInventory inv = make_atlas_inventory();
-    overlay.set_atlas_inventory(&inv);
-    // Both tabs toggled on — the pass viewer is the older surface and
-    // wins; the atlas tab must not paint, and its row count resets.
-    overlay.set_atlas_viewer_visible(true);
-    overlay.set_pass_viewer_enabled(true);
-
-    pulp::canvas::RecordingCanvas canvas;
-    overlay.paint(canvas);
-    REQUIRE(count_text_containing(canvas, "Render Passes") >= 1);
-    REQUIRE(count_text_containing(canvas, "Texture Atlases") == 0);
-    REQUIRE(overlay.atlas_row_count() == 0);
-}
